Refactors propagation routine to base dynamics
Moves the generic propagation routine and initial state validation from rtbp.py to base.py.

This change promotes code reuse and reduces redundancy by centralizing the propagation logic.
